requirements.api.txt
# API Server Dependencies
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0

# Mesh processing for GLB export
trimesh>=4.0.0
scipy>=1.11.0

# Already in main requirements but needed for API
pillow>=10.0.0
numpy>=1.24.0

